[ds-graph] graph_erase_vertice: find erasing vertice in neighbors table and delete
[ds-graph] break down into main, .c, .h files
