+++ graph.c
@@ -1,41 +1,4 @@
-//Implementation of simple directed regular cycled weighted graph
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <stdbool.h>
-#include <glib.h>
-
-#define GRAPH_SIZE 5
-
-typedef struct {
-    int index;
-    int weight;
-} Graph_vertice_neighbor;
-
-typedef struct {
-    int index;
-    GPtrArray *neighbors;
-} Graph_vertice;
-
-typedef struct {
-    GHashTable *vertices_table;
-} Graph;
-
-static int* graph_init_vertices_indexes();
-Graph graph_create();
-void graph_free(Graph **);
-void graph_list(Graph *);
-
-void graph_add_vertice(Graph *, int *);
-bool graph_erase_vertice(Graph *, int);
-Graph_vertice* graph_get_vertice(Graph *, int);
-
-void graph_add_edge(Graph *, int, int, int);
-bool graph_erase_edge(Graph *, int, int);
-static void graph_edge_function_check_pararameters(Graph *, Graph_vertice *, Graph_vertice *, int, int, int);
-static Graph_vertice_neighbor* graph_find_neighbor(GPtrArray *, int);
-static void graph_add_neighbors_table(Graph_vertice *);
-static void graph_add_new_neighbor(GPtrArray *, int, int);
+#include "graph.h"
 
 /*print all vertices indexes in graph*/
 void graph_list(Graph *self) {
@@ -85,7 +48,6 @@ static void graph_add_new_neighbor(GPtrArray *neighbors, int to_vertice, int wei
         fprintf(stderr, "memalloc for Graph_vertice_neighbor failed");
         abort();
     }
-
     new_neighbor->index = to_vertice;
     new_neighbor->weight = weight;
 
@@ -94,7 +56,6 @@ static void graph_add_new_neighbor(GPtrArray *neighbors, int to_vertice, int wei
 
 static void graph_add_neighbors_table(Graph_vertice *self) {
     GPtrArray *ptr_array = g_ptr_array_new_with_free_func(g_free);
-
     self->neighbors = ptr_array;
 }
 
@@ -163,7 +124,6 @@ bool graph_erase_edge(Graph *self, int from_vertice, int to_vertice) {
 
     graph_edge_function_check_pararameters(self, src, dest, from_vertice, to_vertice, stub_weight);
 
-
     if (dest->neighbors != NULL) {
         if (dest->neighbors->len > 0) {
             if(graph_find_neighbor(dest->neighbors, from_vertice) != NULL) {
@@ -209,7 +169,25 @@ bool graph_erase_vertice(Graph* self, int index) {
         abort();
     }
 
-    return g_hash_table_remove(self->vertices_table, (gconstpointer) &index);
+    //erase vertice from neighbors table of other vertices
+    GList* glist = g_hash_table_get_keys(self->vertices_table);
+
+    while(glist != NULL) {
+        int from_vertice = *((int *)(glist->data));
+
+        if(from_vertice != index) {
+            graph_erase_edge(self, from_vertice, index);
+        }
+
+        glist = glist->next;
+    }
+
+    if (self->vertices_table != NULL) {
+        return g_hash_table_remove(self->vertices_table, (gconstpointer) &index);
+    }
+
+    fprintf(stderr, "graph pointer is equal to NULL");
+    abort();
 }
 
 /*
@@ -260,8 +238,10 @@ void graph_free(Graph **self) {
     while(vertices != NULL) {
         vertice = vertices->data;
 
-        g_ptr_array_free(vertice->neighbors, TRUE);
-        vertice->neighbors = NULL;
+        if(vertice->neighbors != NULL) {
+            g_ptr_array_free(vertice->neighbors, TRUE);
+            vertice->neighbors = NULL;
+        }
 
         vertices = vertices->next;
     }
@@ -271,9 +251,10 @@ void graph_free(Graph **self) {
     *self = NULL;
 }
 
-static int* graph_init_vertices_indexes() {
+int* graph_init_vertices_indexes() {
     //static arr also can be used but not so explicit
     int* arr = (int*)malloc(GRAPH_SIZE * sizeof(int));
+
     if (arr == NULL) {
         printf("Memory allocation failed!\n");
         return NULL;
@@ -284,39 +265,4 @@ static int* graph_init_vertices_indexes() {
     }
 
     return arr;
-}
-
-int main() {
-    Graph graph = graph_create();
-    Graph *p_graph = &graph;
-    Graph **p_p_graph = &p_graph;
-
-    int* vertices_indexes = graph_init_vertices_indexes();
-
-    //vertices with indexes from 1 to 5
-    for (int i = 0; i <= GRAPH_SIZE; i++) {
-        graph_add_vertice(p_graph, &vertices_indexes[i]);
-    }
-
-    //0,1,2,4,5
-    graph_erase_vertice(p_graph, 3);
-
-    graph_list(p_graph);
-
-    //pointer to vertice with index 1
-    Graph_vertice *v_1 = graph_get_vertice(p_graph, 1);
-
-    //added 1 -> 2 with weight 4
-    graph_add_edge(p_graph, 1, 2, 4);
-
-    //abort: this graph is not a multigraph
-    // graph_add_edge(p_graph, 2, 1, 4);
-
-    bool is_edge_erased = graph_erase_edge(p_graph, 1, 2);
-
-    graph_list(p_graph);
-
-    graph_free(p_p_graph);
-
-    return 0;
 }
